+++ CubicSpline/CubicSpline_CAPI.h
@@ -8,8 +8,8 @@
 
 extern "C"
 {
-__declspec(dllexport) void __stdcall CSpline_Interp(float *X, float*Y, unsigned int N, float *intXs, float *predYs, unsigned int NPreds);
-__declspec(dllexport) void __stdcall CSpline_Coeffs(float *X, float*Y, unsigned int N, float **coeffs);
+__declspec(dllexport) void __stdcall CSpline_Interp(double *X, double*Y, unsigned int N, double *intXs, double *predYs, unsigned int NPreds);
+__declspec(dllexport) void __stdcall CSpline_Coeffs(double *X, double*Y, unsigned int N, double **coeffs);
 };
